Limit InstanceList join to system_metadata in os-simple-tenant-usage
The simple-tenant-usage API is calling get_active_by_window_joined with
all of the default instance attribute fields when really only
system_metadata is needed (to get the related flavor). That adds
unnecessary load times on the DB query and data processing when
constructing the instance list, especially given this is a query over a
time range of instances, so as you add more instances to scale it's
going to slow down.

This simply limits the expected_attrs in the get_active_by_window_joined
call to system_metadata so we don't load the other unnecessary
attributes.

Closes-Bug: #1383469

Change-Id: Ic77d51cd45730ca664e68cfe0cb1671b5442d4a9