Implement compare-and-swap for instance update
This patch reworks nova.db.sqlalchemy.api._instance_update to remove a
race condition that exists in a critical section of code that
retrieves the instance record from the database, does some checks of
both the new values and existing values of instance fields, and then
tries to update the database with the new values.

We update the value using update_on_match, which does an optimistic,
atomic, lock-free, compare-and-swap on the object. If the update
fails, we do some additional checks to determine the specific error,
which maintains the previous behaviour. There's an exceptionally
small chance of a race when checking for an error if the expected
update conditions were not met when we tried the UPDATE, but were met
when we fetch the instance for error checking. We handle this by
raising a new error, InstanceUpdateConflict.

_instance_update() is simplified by having it return only the updated
instance_ref. instance_update_and_get_original() is updated to fetch
the old value itself before calling _instance_update().

Closes-bug: #1297375
Change-Id: I9cd0f4b620e639b238555983bf6d58deafbaefeb