Make test re-use HTTPRequest part 1
We are using the controller to do test instead of
wsgi URI now, the HTTPRequest is only a fake value
and mostly is not used. Make test class reuse
the Request to remove code complication.
Also, some method set such as 'POST', 'DELETE' etc
are not needed anymore since controller are
directly used.

This patch focus on console_auth_tokens, console_output
and keypairs.

Partially implements blueprint v2-on-v3-api

Change-Id: I359eb78ca0c1dd982e0e63159f6fa66228eaae44