Make test re-use HTTPRequest part 2
We are using the controller to do test instead of
wsgi URI now, the HTTPRequest is only a fake value
and mostly is not used. Make test class reuse
the Request to remove code complication.
Also, some method set such as 'POST', 'DELETE' etc
are not needed anymore since controller are
directly used.

This patch focus on agents, cloudpipe,
cloudpipe_update and certificates.

Partially implements blueprint v2-on-v3-api
Change-Id: I5fe718a5062ad8cc89adf1fd9983f4f1b491361d