Fix py27 gate failure - test_create_instance_both_bdm_formats
Intermittent failure, depending on how the extensions get loaded.
The specific check for HTTPBadRequest is for the exception being
thrown here:
https://github.com/openstack/nova/blob/master/nova/api/openstack/compute/plugins/v3/block_device_mapping_v1.py#L53

To be able to reach this piece of code, both the extensions
'os-block-device-mapping' and 'os-block-device-mapping-v1'
need to be activated. To be sure this happens all the time
we create a both_controllers with an empty blacklist and
use that for this test.

Closes-Bug: #1412547
Change-Id: I284cc9f053909359a8de1a0bd85f61fbfbc98af0