Remove usage of db.service_get_by_args
After the change:

        https://review.openstack.org/#/c/155832

It has become apparent that get_service_by_args method of the DB api
does basically the same thing as the newly introduced
service_get_by_host_and_binary, with several subtle differences. (whether
deleted records are returned, whether disabled services are considered
etc.)

Leaving this as is is likely to lead to confusion, and introduction of
subtle bugs.

The proposed solution is to keep only one. Seeing that the subtleties of
how get_service_by_args works, and the fact that it is really poorly
named, it is deemed better to remove it.

This patch makes all the places that used it now use
get_service_by_host_and_binary instead.

Change-Id: I243236c8feacec5e6edbbbf4cee2ec4e40405bf1