Pass project_id when create networks by os-tenant-networks
os-tenant-networks API is called by normal user. So the create action
should be allowed to create new network which belong to the requester.

This patch passes the project_id down to the nova manager, then make
it create a new network belong to the requester.

The db call network_create_safe is still admin required. That will be
fixed by db permission cleanup patch later.

Change-Id: Iff6ed417a297dcb1594d6871ea18119ce3b3e7b0
Partial-Bug: #1432101