Fix v2.1 list-host to remove 'services' filter
In v2.1, list-host supports 'services' filter also
But this is not case with v2 API. v2 list-host does not support
'services' filter.
v2.1 APIs should be same as v2 in all cases/scenarios.

This is bug fix so not putting as microversions, its just a change to
the v2.1 code so it is equivalent to v2. If later needed then, this can
be reverted back as current ways via microversion.

This was no best test coverage to catch this kind of issue.
This was caught while merging the v2 and v2.1 tests
- Ibd53a7ddd4792942fe2dc45121dcc25fc9864903

Change-Id: I11210babde0c5a668c794cd6aae7eb8c5093a99c
Closes-Bug: #1437184