Remove downgrade support from the cellsv2 api db
The TC approved a cross project spec to stop putting db downgrade into
project migration repos:
https://github.com/openstack/openstack-specs/blob/master/specs/no-downward-sql-migration.rst
This was done after quite a bit of feedback from the operator community,
and everyone was good with it.

This implements that spec for Nova, adjust the tests to pass
accordingly, and changes the test which requires downgrade functions
to one that prohibits them.

Implements bp nova-no-downward-sql-migration

Change-Id: I972e1d79e570d57baac613e56570b49e1b1da28d