devref: add information to clarify nova scope
This adds a document that helps clarify the scope of the Nova mission
statement.

We need to keep the Nova project a manageable size. Agreeing on what
Nova should focus on is a great way to make sure we grow Nova in a
consistent way, leaving open opportunities for other projects to do a
better job of things Nova doesn't see as part of its core mission.

Part of blueprint devref-refresh-liberty

Change-Id: I4e46ee252e6b29b063bcc8204d60d670eb79daef