Catch invalid id input in service_delete
service delete will raise 500 error if the service ID
is invalid. This is caused DBError was raised and nova
didn't handle it. This patch adds check for the exception
and handle it in API layer.

The commit 79a9ca6da345db454f61b82dbd761507141fbddf
clarfy that no microversion is needed.

Closes-Bug: 1471239

Change-Id: Ifde138e72c9a18e01b3f55e188f8d7473ac39133