Remove flavor migration from db_api and nova-manage
This removes the DB API methods for pushing manual migrations of flavor data
to sysmeta. In Liberty, we landed a migration that bans moving forward if this
is not done. Thus, users can't deploy this patch without having made it past
that point, which means they no longer need this.

Related to blueprint drop-flavor-compat

Depends-On: I9daa648f803218bf1d8b7b7686ab73e146eeb04b
Change-Id: Ib5baf62b7bbe7e160d18284140c29bd078d53d3b