Add some devref for AZs
Since the AZ knowledge is mostly tribal and can have some corner cases, we could
help the operators by giving more visibility on how it's made and what to prevent.

The related ticket mentioned below is one example of a common mistake that is
quite not easily fixable from the Nova standpoint since the design is mostly broken.

Change-Id: I092c8caa9e450a68a7a952940b0bb288b8fe6fb0
Related-Bug: #1444841