config options: add hacking check for help text length
Adds a hacking check if a config option provides enough help text.
To do so, the check counts the length of the help text. It uses
a low number in order to avoid a break. As soon as we have agreed
on a higher standard and changed the options accordingly, we can
increase that number.

Change-Id: If31339c428953c6a7bf721ff92e5999e8cb5b569