Add prototype feature classification matrix
Add in feature_classification.ini that makes use of new sphinx
extension feature_matrix. While it is loosely based on the
support_matrix extension, longer term this extension will live
outside the Nova tree. As such, this has been created as a new
separate sphinx plugin.

The matrix has links to wiki page for the CI in the header of the
summary matrix. This is called a target.

Also, there are links to admin docs, API docs, and tempest test uuids
added into the feature details. An option is added to ensure these are
always present in the prototype matrix.

A maturity status is added to be clear about the level of maturity
of each feature. When in maturity mode, this is added into the summary
table in place of the status. There is also some formating for the
different maturity levels.

blueprint feature-classification

Change-Id: Ib5895e8de901f1a282d0f5c0ecb811ff8b451497