Do not use constraints for venv
The venv environment is used by infra scripts to build tarballs and
documents - and used by both check and post jobs. In post jobs, we
currently do not support constraints, so add the non-constraints install
command here so that nova-docs (publishing of documents to
docs.openstack.org) and the tarball jobs work again.

Change-Id: I2e5c0a4d739296b8b9d4ef7f58b69e1eff3c9575