Move placement api request logging to middleware
This change moves the request logging in the placement api to
middleware that is the outermost piece of middleware in the system.

Without this we end up with a situation where some requests which
are not successful appear to be logged with success and other
request do not get logged at all.

By using middleware we assure that the logging of the beginning of
the request and _any_ exit of the request will be logged because the
middleware can be the first and last thing the request interacts
with.

Change-Id: I4215cc69cedae5637102b75e0b54fd26acb1826c
Closes-Bug: #1619690