[placement] Correct serialization of inventory collections
The correct form is for resource_provider_generation to be its on
separate key, not repeated in each individual inventory. This was
probably caused by the refactoring that create _send_inventory and
_send_inventories and a lack of sufficient test coverage.

Fixing this involves changes in both the placement api service, and
in the (thus far) only existing client, the scheduler reporting
client used in the resource tracker. The reporting client can be a
bit simpler now because of the cleaner behavior in the api.

Tests have been updated accordingly.

Change-Id: I3af1c7686a45c1a0d70fe704d3c7938810eff6a3
Closes-Bug: #1619771