[placement] Manage log and other output in gabbi fixure
Recent changes to the placement API added more verbose logging and
output across both stdout and stderr. In test runs this resulted in
a combination of far too much output as well as output that was
impossible to read because it was interleaved.

This change uses fixtures from nova.tests.fixtures to capture
stray output streams and logs.

It also adds a specific logger for the ERROR level that overrides
the fixture log handling to send that level to the default stderr.
This is required in order to effectively see error messages on both
side of the wsgi-intercept used when running gabbi. Without this,
the errors are swallowed and test development and debugging,
especially locally for TDD, is hamstrung.

Closes-Bug: #1623573
Change-Id: I300456b0261dfc0ae46c799e5a03cbeab25106c2