[placement] fix typo in call to create auth middleware
The typo meant that the intended non global config was not being
used, and instead the config was loaded in the normal way. This
worked because the config loaded by the project name is the same
as the config loaded otherwise.

There's no test coverage of this in unit and functional tests
because the NoAuthMiddleware is always used based on overridden
configuration. The gate tests that turn on placement will exercise
this code.

Change-Id: I9f1b8905fd9d5aadac61718e24c76b1b2f80410d
Closes-Bug: #1645306