doc: code review considerations for online data migrations
Online data migrations that move something from the cell database
to the API database typically create new things in the API DB and
on a query, look in the API DB first and then fallback to the cell
database. We started supporting multiple cells in Pike, so any online
data migrations that move things to the API DB after that will need
to be multi-cell aware for the fallback lookup code.

This just adds a reminder in the code review guide.

Change-Id: If0d7d9b80e336b696aaf87ec13ac18daa1068357