Add fault-filling into instance_get_all_by_filters_sort()
For some legacy reason we've never been able to request fault in DB
queries for instances, and have done it later in a batch operation
from the API if needed. So that we can keep this all to one operation
per cell, this patch adds 'fault' as a valid "manual join" column
to the query call.

Note that this is terribly inefficient, just like the rest of our
so-called "manual join" operations. However, because of the way the
fault query method works, we don't actually iterate the full instance
list any more times to do this (although we _do_ perform another DB
query to get the list of faults).

We really should work at optimizing those manual joins at some point,
but that is a whole separate effort on its own and this doesn't really
make the situation any worse than it is now.

Change-Id: Ia99ea54d74b60c6d2c95f0fa531456b0e7b64200