Follow up on removing old-style quotas code
This is a follow up to another patch [1] that removed the old-style
quotas code which is no longer in use. Here, we remove the 'reserved'
key from quotas internally and set it in the APIs where it's expected.
We also remove code that accesses the quota_usages and reservations
tables as they're no longer used.

[1] https://review.openstack.org/#/c/511689

Change-Id: I75291571468ddb79b7561810de0953bb462548e3