doc: link in some Sydney summit content
This adds some links to talks from the Sydney summit to the docs
for cells v2, bug triage, and the metadata service.

While adding a "References" section to the metadata docs, I figured
it was also useful to link to a blog post from mikal about vendordata
since it also includes code samples.

Change-Id: Ifc47a5472db37f5526004d2e00751365a026975a